pyproject.toml
[project]
name = "ad_moderation"
version = "1.0.0"
description = "Ad Moderation OpenEnv Benchmark"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openenv>=0.1.13",
    "openai",
    "python-dotenv",
]

[project.scripts]
server = "server.app:main"

[tool.openenv]
name = "AdModeration"
version = "1.0.0"
tasks = [
    "spam_detection",
    "hate_speech_detection",
    "misinformation_detection",
    "safe_content"
]
entry_point = "server:app"
